Refactoring HyperNEATVisualizationUtil #258
